+++ src/mbot.h
@@ -3,7 +3,6 @@
 
 #include <pico/stdlib.h>
 #include <pico/mutex.h>
-#include <pico/multicore.h>
 #include <pico/time.h>
 #include <hardware/gpio.h>
 #include <mbot/motor/motor.h>
@@ -15,19 +14,18 @@
 #include <mbot/imu/imu.h>
 #include <rc/math/filter.h>
 #include <rc/mpu/mpu.h>
-#include <comms/common.h>
-#include <comms/protocol.h>
-#include <comms/listener.h>
-#include <comms/topic_data.h>
-#include <mbot_lcm_msgs_serial.h>
+
+// #include <comms/common.h>
+// #include <comms/protocol.h>
+// #include <comms/listener.h>
+// #include <comms/topic_data.h>
+// #include <mbot_lcm_msgs_serial.h>
 
 #include <math.h>
 #include <inttypes.h>
 
-#include "mbot_channels.h"
-
-
-#define MESSAGE_CONFIRMATION_CHANNEL "MSG_CONFIRM"
+#include "mbot_comms.h"
+// #include "mbot_channels.h"
 
 // TODO: Decide which controller is used, open loop = 1, PID = 0
 #define OPEN_LOOP 1
@@ -51,31 +49,7 @@ enum drive_modes{
     MODE_MBOT_VEL = 3
 };
 
-/*
-* Messages used by the MBot code, 
-* we also use these to store state
-*/
-// origin: mbot
-serial_mbot_imu_t mbot_imu = {0};
-serial_pose2D_t mbot_odometry = {0};
-serial_mbot_encoders_t mbot_encoders = {0};
-serial_twist2D_t mbot_vel = {0};
-serial_mbot_motor_pwm_t mbot_motor_pwm = {0};
-serial_mbot_motor_vel_t mbot_motor_vel = {0};
 
-// origin: comms
-serial_twist2D_t mbot_vel_cmd = {0};
-serial_mbot_motor_pwm_t mbot_motor_pwm_cmd = {0};
-serial_mbot_motor_vel_t mbot_motor_vel_cmd = {0};
-serial_timestamp_t mbot_received_time = {0};
-
-//callback functions
-void timestamp_cb(serial_timestamp_t *msg);
-void reset_encoders_cb(serial_mbot_encoders_t *msg);
-void reset_odometry_cb(serial_pose2D_t *msg);
-void mbot_vel_cmd_cb(serial_twist2D_t *msg);
-void mbot_motor_vel_cmd_cb(serial_mbot_motor_vel_t *msg);
-void mbot_motor_pwm_cmd_cb(serial_mbot_motor_pwm_t *msg);
 bool mbot_loop(repeating_timer_t *rt);
 void mbot_read_encoders(serial_mbot_encoders_t* encoders);
 void mbot_calculate_motor_vel(serial_mbot_encoders_t encoders, serial_mbot_motor_vel_t *motor_vel);
