+++ src/mbot_comms.c
@@ -1,0 +1,48 @@
+#include "mbot_comms.h"
+
+serial_mbot_imu_t mbot_imu = {0};
+serial_pose2D_t mbot_odometry = {0};
+serial_mbot_encoders_t mbot_encoders = {0};
+serial_twist2D_t mbot_vel = {0};
+serial_mbot_motor_pwm_t mbot_motor_pwm = {0};
+serial_mbot_motor_vel_t mbot_motor_vel = {0};
+
+serial_twist2D_t mbot_vel_cmd = {0};
+serial_mbot_motor_pwm_t mbot_motor_pwm_cmd = {0};
+serial_mbot_motor_vel_t mbot_motor_vel_cmd = {0};
+serial_timestamp_t mbot_received_time = {0};
+
+void register_topics()
+{
+    // Subscriptions
+    comms_register_topic(MBOT_TIMESYNC, sizeof(serial_timestamp_t), (Deserialize)&timestamp_t_deserialize, (Serialize)&timestamp_t_serialize, (MsgCb)&timestamp_cb);
+    comms_register_topic(MBOT_ODOMETRY_RESET,  sizeof(serial_pose2D_t), (Deserialize)&pose2D_t_deserialize, (Serialize)&pose2D_t_serialize, (MsgCb)&reset_odometry_cb);
+    comms_register_topic(MBOT_ENCODERS_RESET, sizeof(serial_mbot_encoders_t), (Deserialize)&mbot_encoders_t_deserialize, (Serialize)&mbot_encoders_t_serialize, (MsgCb)&reset_encoders_cb);
+    comms_register_topic(MBOT_MOTOR_PWM_CMD, sizeof(serial_mbot_motor_pwm_t), (Deserialize)&mbot_motor_pwm_t_deserialize, (Serialize)&mbot_motor_pwm_t_serialize, (MsgCb)mbot_motor_vel_cmd_cb);
+    comms_register_topic(MBOT_MOTOR_VEL_CMD, sizeof(serial_mbot_motor_vel_t), (Deserialize)&mbot_motor_vel_t_deserialize, (Serialize)&mbot_motor_vel_t_serialize, (MsgCb)mbot_motor_pwm_cmd_cb);
+    comms_register_topic(MBOT_VEL_CMD, sizeof(serial_twist2D_t), (Deserialize)&twist2D_t_deserialize, (Serialize)&twist2D_t_serialize, (MsgCb)mbot_vel_cmd_cb);
+
+    // Published Topics
+    comms_register_topic(MBOT_ODOMETRY, sizeof(serial_pose2D_t), (Deserialize)&pose2D_t_deserialize, (Serialize)&pose2D_t_serialize, NULL);
+    comms_register_topic(MBOT_IMU, sizeof(serial_mbot_imu_t), (Deserialize)&mbot_imu_t_deserialize, (Serialize)&mbot_imu_t_serialize, NULL);
+    comms_register_topic(MBOT_ENCODERS, sizeof(serial_mbot_encoders_t), (Deserialize)&mbot_encoders_t_deserialize, (Serialize)&mbot_encoders_t_serialize, NULL);
+    comms_register_topic(MBOT_VEL, sizeof(serial_twist2D_t), (Deserialize)&twist2D_t_deserialize, (Serialize)&twist2D_t_serialize, NULL);
+    comms_register_topic(MBOT_MOTOR_VEL, sizeof(serial_mbot_motor_vel_t), (Deserialize)&mbot_motor_vel_t_deserialize, (Serialize)&mbot_motor_vel_t_serialize, NULL);
+    comms_register_topic(MBOT_MOTOR_PWM, sizeof(serial_mbot_motor_pwm_t), (Deserialize)&mbot_motor_pwm_t_deserialize, (Serialize)&mbot_motor_pwm_t_serialize, NULL);
+
+}
+
+int mbot_init_comms(void){
+    printf("Initializing LCM serial communication...\r\n");
+    comms_init_protocol();
+    comms_init_topic_data();
+    register_topics();
+
+    // launch the other core's comm loop
+    printf("starting comms on core 1...\r\n");
+    multicore_launch_core1(comms_listener_loop);
+
+    // wait for other core to get rolling
+    sleep_ms(500);
+    return MBOT_OK;
+}
